backend/fastapi_service/requirements.txt
# Python Version: 3.10.11

fastapi==0.115.0
uvicorn==0.30.0

# Imaging
pillow==12.0.0

# HTTP requests
requests==2.32.4

# Validation
pydantic==2.11.7
pydantic_core==2.33.2

# File upload support
python-multipart==0.0.9

# Environment variable support
python-dotenv==1.2.1
pydantic_settings==2.12.0

# Google Generative AI SDK
google-generative-ai==1.52.0